+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.0 (November 16, 2020)
+
+Added Jobs:
+
+* b/collection/coalesce
+* b/collection/group
+
 # 1.0.0 (November 5th, 2020)
 
 Initial version publication.